Move filterDisabledOption to module
This was a remaining method that could be safely
moved to the permissions module. All usages have
been updated.
