+++ src/lib/options.js
@@ -2,7 +2,7 @@ const { castArray, sortBy } = require('lodash')
 
 const config = require('../../config')
 const authorisedRequest = require('../lib/authorised-request')
-const { filterDisabledOption } = require('../apps/filters')
+const { filterDisabledOption } = require('../modules/permissions/filters')
 const { transformObjectToOption } = require('../apps/transformers')
 
 async function getOptions (token, key, { createdOn, currentValue, includeDisabled = false, sorted = true, term, id, queryString = '', context } = {}) {
