+++ src/modules/permissions/filters.js
@@ -1,5 +1,20 @@
 const { get, intersection } = require('lodash')
 
+function filterDisabledOption ({ currentValue = null, createdOn } = {}) {
+  const createdOnTime = Date.parse(createdOn) || Date.now()
+
+  return function (item) {
+    if (!item.disabled_on) {
+      return true
+    }
+
+    const isDisabled = Date.parse(item.disabled_on) > createdOnTime
+    const isCurrentValue = item.id === currentValue
+
+    return (isDisabled || isCurrentValue)
+  }
+}
+
 function filterNonPermittedItem (userPermissions) {
   return function (item) {
     const hasPermissions = get(item, 'permissions.length', 0)
@@ -12,4 +27,5 @@ function filterNonPermittedItem (userPermissions) {
 
 module.exports = {
   filterNonPermittedItem,
+  filterDisabledOption,
 }
